math_service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
sqlalchemy==2.0.30
aiosqlite==0.19.0
python-dotenv==1.0.1  # manage API keys via .env file